Revise pytests, and extract pytests for references related processes into its own test module 'test_filehashstore_references'
