+++ tests/test_filehashstore_interface.py
@@ -542,6 +542,22 @@ def test_tag_object_pid_refs_file(pids, store):
         assert os.path.exists(pid_refs_file_path)
 
 
+def test_tag_object_pid_refs_file_exists(pids, store):
+    """Test tag object throws exception when pid refs file already exists."""
+    test_dir = "tests/testdata/"
+    for pid in pids.keys():
+        path = test_dir + pid.replace("/", "_")
+        object_metadata = store.store_object(pid, path)
+        cid = object_metadata.id
+        store.tag_object(pid, cid)
+        pid_refs_file_path = store.get_refs_abs_path("pid", pid)
+        assert os.path.exists(pid_refs_file_path)
+        cid_refs_file_path = store.get_refs_abs_path("cid", cid)
+        assert os.path.exists(cid_refs_file_path)
+        with pytest.raises(FileExistsError):
+            store.tag_object(pid, cid)
+
+
 def test_tag_object_pid_refs_file_content(pids, store):
     """Test tag object creates the pid reference file contains the correct cid."""
     test_dir = "tests/testdata/"
@@ -580,7 +596,7 @@ def test_tag_object_cid_refs_file_content(pids, store):
         assert pid_refs_cid == pid
 
 
-def test_tag_object_with_existing_cid_refs_file(pids, store):
+def test_tag_object_cid_refs_file_exists(pids, store):
     """Test tag object raises exception when trying to add another cid to an
     existing pid reference file and that a cid reference file is not created."""
     test_dir = "tests/testdata/"
